# docs/swagger.yaml
definitions:
  util.ApiResp:
    properties:
      code:
        type: integer
      data:
        type: object
      msg:
        type: string
      status:
        type: boolean
    type: object
info:
  contact: {}
  license: {}
paths:
  /admin/login:
    post:
      consumes:
      - multipart/form-data
      - application/x-www-form-urlencoded
      description: 统一登录接口，暂时接收账号密码登录和手机验证码登录
      parameters:
      - description: 登录类型1:用户名密码登录，2手机号验证码登录
        in: formData
        name: type
        required: true
        type: integer
      - description: 用户真实姓名
        in: formData
        name: username
        type: string
      - description: 密码
        in: formData
        name: password
        type: string
      - description: 手机号
        in: formData
        name: phone
        type: string
      - description: 手机验证码
        in: formData
        name: captcha
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.ApiResp'
        "302":
          description: Found
          schema:
            $ref: '#/definitions/util.ApiResp'
      summary: 登录
      tags:
      - 用户信息
  /api/captcha:
    get:
      consumes:
      - multipart/form-data
      - application/x-www-form-urlencoded
      description: 获取图形验证码图片
      parameters:
      - description: 手机号
        in: query
        name: phone
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.ApiResp'
        "302":
          description: Found
          schema:
            $ref: '#/definitions/util.ApiResp'
      summary: 图形验证码
      tags:
      - 公共接口
  /api/sms:
    post:
      consumes:
      - multipart/form-data
      - application/x-www-form-urlencoded
      description: 发送短信接口
      parameters:
      - description: 手机号
        in: formData
        name: phone
        required: true
        type: string
      - description: 图形验证码的值
        in: formData
        name: captcha
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.ApiResp'
        "302":
          description: Found
          schema:
            $ref: '#/definitions/util.ApiResp'
      summary: 短信接口
      tags:
      - 公共接口
swagger: "2.0"